other: use f32 for process percentage values (#1212)
* other: use f32 for process percentage values

This cuts down memory by a tiny bit, and we don't need a full f64 for
percentage values.

* fix for macos and windows

+++ src/app/data_harvester/processes.rs
@@ -45,10 +45,10 @@ pub struct ProcessHarvest {
     pub parent_pid: Option<Pid>,
 
     /// CPU usage as a percentage.
-    pub cpu_usage_percent: f64,
+    pub cpu_usage_percent: f32,
 
     /// Memory usage as a percentage.
-    pub mem_usage_percent: f64,
+    pub mem_usage_percent: f32,
 
     /// Memory usage as bytes.
     pub mem_usage_bytes: u64,
